fmm
An implementation of the Fast Multipole Method. Takes an n-vector of
sources, an n-vector of targets, an n-vector of points to evaluate at,
and the degree of precision of the output. Both the sources and targets
must be within the interval [0,1]. The output is an approximation to the
resulting n-vector when the Cauchy matrix created by the sources and
targets is multiplied with the n-vector of points. Runs in O(nlogn)
time, compared to a naive multiplication, which runs in  O(n) time.